backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
pydantic-settings==2.10.1
python-multipart==0.0.20
pytest==8.4.1
httpx==0.28.1
passlib==1.7.4
PyJWT==2.10.1
